Zinsberechnung korrigiert

+++ src/application/accounting/AccountingEntry.java
@@ -1,11 +1,13 @@
 package application.accounting;
 
+import java.math.BigDecimal;
+
 public class AccountingEntry {
     
     private int tag; //[1-360]
-    private long betrag; //in 100tel cent
+    private BigDecimal betrag; //in 100tel cent
     
-    public AccountingEntry(int tag, long betrag){
+    public AccountingEntry(int tag, BigDecimal betrag){
         this.tag = tag;
         this.betrag = betrag;
     }
@@ -14,7 +16,7 @@ public class AccountingEntry {
         return this.tag;
     }
     
-    public long getBetrag() {
+    public BigDecimal getBetrag() {
         return this.betrag;
     }
 }
